kafka_broker/producer/requirements.txt
kafka-python>=2.0.2
kafka-python[snappy]
websockets>=12.0